Use a timeout for dashboard server
When the timeout is exceeded, a TimoutException is thrown causing the service
to fail.

All dashboard services return (almost) immediately, so actions do not really
make sense here. The only exception is when there is a problem with the dasboard
connection, which is why we introduce the timeout.

This way, service calls will not block forever, when connection to the
dashboard server got lost or if the server isn't answering due to any other
reason.